requirements.txt
wheel==0.34.2
requests==2.22.0
Flask==1.1.2
Flask-Login==0.4.1
Flask-WTF==0.14.3
Flask-SQLAlchemy==2.4.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-gzip==0.2
psycopg2==2.8.3
sh==1.12.14
boto3==1.9.170
python-dotenv==0.13.0
redis==3.5.3
pycrypto==2.6.1
